// frontend/src/api/generated/models/player-spdetail.ts
/* tslint:disable */
 
/**
 * GESTALOKA API
 * マルチプレイ・テキストMMO - ゲスタロカ API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * プレイヤーSP詳細情報
 * @export
 * @interface PlayerSPDetail
 */
export interface PlayerSPDetail {
    /**
     * 
     * @type {number}
     * @memberof PlayerSPDetail
     */
    'user_id': number;
    /**
     * 
     * @type {string}
     * @memberof PlayerSPDetail
     */
    'username': string;
    /**
     * 
     * @type {string}
     * @memberof PlayerSPDetail
     */
    'email': string;
    /**
     * 
     * @type {number}
     * @memberof PlayerSPDetail
     */
    'current_sp': number;
    /**
     * 
     * @type {number}
     * @memberof PlayerSPDetail
     */
    'total_earned': number;
    /**
     * 
     * @type {number}
     * @memberof PlayerSPDetail
     */
    'total_consumed': number;
    /**
     * 
     * @type {Date}
     * @memberof PlayerSPDetail
     */
    'last_daily_recovery': Date | null;
    /**
     * 
     * @type {number}
     * @memberof PlayerSPDetail
     */
    'consecutive_login_days': number;
    /**
     * 
     * @type {Date}
     * @memberof PlayerSPDetail
     */
    'created_at': Date;
    /**
     * 
     * @type {Date}
     * @memberof PlayerSPDetail
     */
    'updated_at': Date;
}